+++ component/include/pl_event_handler.h
@@ -12,7 +12,10 @@ namespace PL {
 template <class Source, class... Args>
 class EventHandler {
 public:
+  EventHandler() = default;
   virtual ~EventHandler() = default;
+  EventHandler(const EventHandler&) = delete;
+  EventHandler& operator=(const EventHandler&) = delete;
 
   /// @brief Event handler callback method
   /// @param source event source